requirements.txt
supabase==2.5.0
google-genai==0.3.0
python-dotenv==1.0.1
pytz==2024.1
tenacity==8.2.3
pydantic==2.6.0
pytest==8.0.0
pytest-asyncio==0.23.3